Move Weather to containers and update index files
