+++ src/components/index.js
@@ -1,9 +1,10 @@
-import WeatherIcon from './WeatherIcon';
-import ForecastBlock from './ForecastBlock';
-import Temperature from './Temperature';
-
-import HourlyForecast from './HourlyForecast';
 import CurrentForecast from './CurrentForecast';
+import DailyForecast from './DailyForecast';
+import ForecastBlock from './ForecastBlock';
+import Hour from './Hour';
+import HourlyForecast from './HourlyForecast';
+import Temperature from './Temperature';
+import WeatherIcon from './WeatherIcon';
 
 export {
   WeatherIcon,
@@ -11,4 +12,6 @@ export {
   ForecastBlock,
   CurrentForecast,
   HourlyForecast,
+  DailyForecast,
+  Hour,
 };
